delete user working ok

+++ controllers/userController.js
@@ -71,10 +71,12 @@ async function update(req, res) {}
 
 // Remove the specified resource from storage.
 async function destroy(req, res) {
+  if (req.user.id !== req.params.id) return res.redirect("back");
   await Tweet.deleteMany({ user: req.params.id });
   await User.updateMany({}, { $pull: { following: req.params.id } });
   await User.updateMany({}, { $pull: { followers: req.params.id } });
   await User.findByIdAndDelete(req.params.id);
+  res.redirect("/logout");
 }
 
 async function logout(req, res) {
